Create zooanimals
The code above was a project I did for sorting zoo animals depending on specific categories. An example would be having separate commands for animal births and deaths and transfers to different zoos and whatnot. The code also sorts the lists of animals alphabetically when animals are either added or removed.